version 9.0.0b12
Added/fixed the logger.
Added a Printer class to Utils so messages are formated and printed to
the console.
Added an Inputs class that will write the control file.
Fixed a bug in the topo averaging that caused the model to crash.
Zeros will be added by default to input files that have data elements
that are blank, NA, or ' '.
Clarified some code comments.