v9.0.0b20
Fixed a critical bug in the way inflow and met input data was being
read. Bug would duplicate columns because the Temperature and Flow
headers were the same name. This translated to duplicate keys in the
dictionary.

Removed debugging info from the logging file.

Changed 'dir' to 'tran' since that is a python reserved name.

Removed methods that called arcpy so it's not included when building
executables.