Updated calculation of extinction coefficient
Updated calculation of extinction coefficient to be consistent with Chen 1998, Norman and Welles 1983 when using canopy cover.

When using LAI, modifed equation to calculate extinction coefficient as a function of the sample path length instead of directly calculating fraction passed using the LAI data.

Added code to recognize a situation when the incoming solar beam enters the top of the canopy instead of the side. The path length is shorter in this case and requires use of the relative vegetation height in order to determine the path length.