Add sheetname and datetime types
Add dictionary to hold the xlsx sheet names for each input file.
Updated the dtype dictionary so datetime values are "datetime". This was needed to handle reading and checking of values from both csv and excel files. Excel files are read as a true datetime type while the values are strings in csv files.